+++ stochpylib/__init__.py
@@ -1,5 +1,17 @@
-from stochpylib import distributions, montecarlo, probability, timeseries
+from stochpylib import (
+    distributions,
+    gaussian_processes,
+    montecarlo,
+    probability,
+    timeseries,
+)
 
-__all__ = ["distributions", "montecarlo", "probability", "timeseries"]
+__all__ = [
+    "distributions",
+    "gaussian_processes",
+    "montecarlo",
+    "probability",
+    "timeseries",
+]
 
 __version__ = "0.2.0"